separate out tokenCache() class to separate file
