+++ app.js
@@ -4,7 +4,10 @@ import got from "got"
 import express from "express"
 import axios from "axios"
 import cors from "cors"
+
+import tokenCache from "./tokenCache.js"
 // import { ApolloClient, gql, InMemoryCache, ApolloProvider } from '@apollo/client';
+
 dotenv.config()
 
 const FFLOGS_CLIENT_ID = process.env.FFLOGS_CLIENT_ID;
@@ -40,133 +43,6 @@ const TWITCH_OPTS = {
   }
 }
 
-class fflogsTokenCache {
-    constructor() {
-        this.tokenPromise = null;
-        this.timer = null;
-        // get the first token
-        this._getNewToken().catch(err => {
-            console.log("Error fetching initial FFLogs token", err);
-        });
-    }
-
-    getToken() {
-        if (this.tokenPromise) {
-            return this.tokenPromise.then(tokenData => {
-                // if token has expired
-                if (tokenData.expires < Date.now()) {
-                    console.log('refreshing fflogs token');
-                    return this._getNewToken();
-                } else {
-                    // console.log(`returning token: ${tokenData.token}`);
-                    return tokenData.token;
-                }
-            });
-        } else {
-            return this._getNewToken();
-        }
-    }
-
-    // non-public method for getting a new token
-    _getNewToken() {
-        this.tokenPromise = got(FFLOGS_AUTH, FFLOGS_OPTS).then(token => {
-            // make resolve value be an object that contains the token and the expiration
-            // set timer to get a new token automatically right before expiration
-            var accessToken = JSON.parse(token["body"])["access_token"];
-            var tokenExpiration = JSON.parse(token["body"])["expires_in"];
-            var tokenBeforeTime = 300000; // 5 min in ms
-            console.log(`\nfflogs\n\naccessToken:\n${accessToken}\n\ntokenExpiration:\n${Date.now() + tokenExpiration}\n`)
-            this._scheduleTokenRefresh(tokenExpiration - tokenBeforeTime);
-            return {
-                token: accessToken,
-                expires: Date.now() + tokenExpiration
-            }
-        }).catch(err => {
-            // up error, clear the cached promise, log the error, keep the promise rejected
-            console.log(err);
-            this.tokenPromise = null;
-            throw err;
-        });
-        return this.tokenPromise;
-    }
-
-    // schedule a call to refresh the token before it expires
-    _scheduleTokenRefresh(t) {
-        if (this.timer) {
-            clearTimeout(this.timer);
-        }
-        this.timer = setTimeout(() => {
-            this._getNewToken().catch(err => {
-                console.log("Error updating FFLogs token before expiration", err);
-            });
-            this.timer = null;
-        }, t);
-    }
-}
-
-class twitchTokenCache {
-  constructor() {
-    this.tokenPromise = null;
-    this.timer = null;
-    // get the first token
-    this._getNewToken().catch(err => {
-      console.log("Error fetching initial Twitch token", err);
-    });
-  }
-
-  getToken() {
-    if (this.tokenPromise) {
-      return this.tokenPromise.then(tokenData => {
-        // if token has expired
-        if (tokenData.expires < Date.now()) {
-          console.log('refreshing twitch token');
-          return this._getNewToken();
-        } else {
-          return tokenData.token;
-        }
-      });
-    } else {
-      return this._getNewToken();
-    }
-  }
-
-  // non-public method for getting a new token
-  _getNewToken() {
-    this.tokenPromise = got(TWITCH_AUTH, TWITCH_OPTS).then(token => {
-      // make resolve value be an object that contains the token and the expiration
-      // set timer to get a new token automatically right before expiration
-      var accessToken = JSON.parse(token["body"])["access_token"];
-      var tokenExpiration = JSON.parse(token["body"])["expires_in"];
-      var tokenBeforeTime = 300000; // 5 min in ms
-      console.log(`\ntwitch\n\naccessToken:\n${accessToken}\n\ntokenExpiration:\n${Date.now() + tokenExpiration}\n`)
-      this._scheduleTokenRefresh(tokenExpiration - tokenBeforeTime);
-      return {
-        token: accessToken,
-        expires: Date.now() + tokenExpiration
-      }
-    }).catch(err => {
-      // up error, clear the cached promise, log the error, keep the promise rejected
-      console.log(err);
-      this.tokenPromise = null;
-      throw err;
-    });
-    return this.tokenPromise;
-  }
-
-  // schedule a call to refresh the token before it expires
-  _scheduleTokenRefresh(t) {
-    if (this.timer) {
-      clearTimeout(this.timer);
-    }
-    this.timer = setTimeout(() => {
-      this._getNewToken().catch(err => {
-        console.log("Error updating Twiich token before exipration", err);
-      });
-      this.timer = null;
-    }, t);
-  }
-}
-
 console.log(FFLOGS_CLIENT_ID + ":" + FFLOGS_CLIENT_SECRET);
 console.log(TWITCH_CLIENT_ID + ":" + TWITCH_CLIENT_SECRET);
 
@@ -179,34 +55,8 @@ const server = https.createServer((req, res) => {
   res.end('I did it!');
 });
 
-var fflogsToken = new fflogsTokenCache();
-var twitchToken = new twitchTokenCache();
-
-// fflogsToken.getToken().then(token => {
-//   // console.log(token)
-//   const options = {
-//     method: "GET",
-//     searchParams: {query: query},
-//     headers: {
-//       "Authorization": `Bearer ${token}`
-//     },
-//     retry: {
-//       limit: 2,
-//       statusCodes: [401],
-//       errorCodes: ['ERR_GOT_REQUEST_ERROR']
-//     }
-//   };
-
-//   const data = got("https://www.fflogs.com/api/v2/client", options);
-//   return data;
-// }).then(data => {
-//   var data = JSON.parse(data["body"]);
-//   console.log(JSON.stringify(data, null, 2));
-// });
-
-// server.listen(port, hostname, () => {
-//   console.log(`Server running at http://${hostname}:${port}/`);
-// });
+var fflogsToken = new tokenCache('fflogs', FFLOGS_AUTH, FFLOGS_OPTS);
+var twitchToken = new tokenCache('twitch', TWITCH_AUTH, TWITCH_OPTS);
 
 var app = express();
